+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Bingo!".
 - Tests for "Duck hunt".
 - Tests for "Condition overshadowing".
+- Tests for "Equalizing arrays".
 
 ## [1.15.0] - 2022-11-30
 ### Added